+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### 0.0.x (July 25, 2017)
 
   * Updated the README.md file,
+  * Added the method replaceWith(),
   * ...,